backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
sqlmodel==0.0.22
aiofiles==23.2.1
python-dotenv==1.0.1
gitpython==3.1.43
PyGithub==2.5.0
requests==2.32.3
starlette==0.38.5
sse-starlette==2.0.0
watchfiles==0.22.0
pytz==2024.1
